+++ src/python/magnum/math.vectorintegral.cpp
@@ -98,10 +98,13 @@ template<class T> void vectorsIntegral(py::module_& m, py::class_<Math::Vector2<
     vector4<T>(vector4_);
 }
 
-template<class T> void vectorsIntegralSigned(py::class_<Math::Vector2<T>>& vector2_, py::class_<Math::Vector3<T>>& vector3_, py::class_<Math::Vector4<T>>& vector4_) {
+template<class T> void vectorsIntegralSigned(py::module_& m, py::class_<Math::Vector2<T>>& vector2_, py::class_<Math::Vector3<T>>& vector3_, py::class_<Math::Vector4<T>>& vector4_) {
     everyVectorSigned(vector2_);
     everyVectorSigned(vector3_);
     everyVectorSigned(vector4_);
+    vectorSigned(m, vector2_);
+    vectorSigned(m, vector3_);
+    vectorSigned(m, vector4_);
     vector2Signed<T>(vector2_);
 }
 
@@ -138,7 +141,7 @@ void mathVectorIntegral(py::module_& root, py::module_& m) {
 
     /* Now register the generic from-list constructors and everything else */
     vectorsIntegral<Int>(m, vector2i, vector3i, vector4i);
-    vectorsIntegralSigned<Int>(vector2i, vector3i, vector4i);
+    vectorsIntegralSigned<Int>(m, vector2i, vector3i, vector4i);
     vectorsIntegral<UnsignedInt>(m, vector2ui, vector3ui, vector4ui);
 }